Add files via upload
This is the dataset from Kaggle and its original website is https://www.kaggle.com/mandygu/lingspam-dataset.

Context
The Ling-Spam dataset is a collection of 2,893 spam and non-spam messages curated from the Linguist List. These messages focus on linguistic interests around job postings, research opportunities and software discussion.

Content
Header information is removed.
2412 legitimate messages
481 spam mesissages

Acknowledgements
All acknowledgements go to the original authors of A Memory-Based Approach to Anti-Spam Filtering for Mailing Lists. The dataset was made publicly available as a part of that paper. The raw data, which is partitioned into subdirectories, is hosted here.